==> part_iii_machine_learning_application/scripts_final/03_test_kernel_PCA.ipynb <==
## In dieser Datei wurde das verwenden einer PCA gestestet, dies wurde letztens Endes jedoch verworfen

import pandas as pd
import numpy as np
import os
import glob
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import KernelPCA
from sklearn.decomposition import TruncatedSVD
import matplotlib.pyplot as plt
import seaborn as sns

pathToMaschenFile = os.path.join("..", "..", "data", "maschenPickled.pickle")
dfMaschen = pd.read_pickle(pathToMaschenFile)
dfMaschen

Kernel PCA Analysis

dfMaschen.drop(dfMaschen.tail(513000).index,inplace=True)
dfMaschen

#Standardize dataframe
features = ['latitude', 'longitude', 'movement_state', 'GNSS_velocity']# Separating out the features
standMasch_df = dfMaschen.loc[:, features].values# Separating out the target
standMasch_df = StandardScaler().fit_transform(standMasch_df)

standMasch_df

#apply kernel and then perform PCA
kernel_pca = KernelPCA(n_components=2, kernel="rbf", gamma=None, fit_inverse_transform=True, alpha=0.1)
principalComponents = kernel_pca.fit_transform(standMasch_df)

#Store results
principalDf = pd.DataFrame(data = principalComponents, columns = ['principal component 1', 'principal component 2'])

#Plot results
fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('Principal Component 1', fontsize = 15)
ax.set_ylabel('Principal Component 2', fontsize = 15)
ax.set_title('2 component kernel PCA', fontsize = 20)

ax.scatter(principalDf['principal component 1'],principalDf['principal component 2'])
ax.grid()

#Extract first cluster (with high principal component 2)
cluster1index = principalDf.loc[principalDf['principal component 2']>=0.25].index
cluster1df = dfMaschen.loc[cluster1index]

cluster1df.describe()

Kernel PCA with 3 Components

#kernel PCA with 3 components
kernel_pca = KernelPCA(n_components=3, kernel="rbf", gamma=None, fit_inverse_transform=True, alpha=0.1)
principalComponents_3 = kernel_pca.fit_transform(standMasch_df)

#Store results
principalDf_3components = pd.DataFrame(data = principalComponents_3, columns = ['principal component 1', 'principal component 2', 'principal component 3'])

#Plot results
fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(projection='3d') 
ax.set_xlabel('Principal Component 1', fontsize = 15)
ax.set_ylabel('Principal Component 2', fontsize = 15)
ax.set_zlabel('Principal Component 3', fontsize = 15)
ax.set_title('3 Component Kernel PCA', fontsize = 20)

ax.scatter(principalDf_3components['principal component 1'],principalDf_3components['principal component 2'],principalDf_3components['principal component 3'])
plt.show()

# Linear Kernel PCA

#apply linear kernel and then perform PCA
kernel_pca = KernelPCA(n_components=2, kernel="linear", gamma=None, fit_inverse_transform=True, alpha=0.1)
principalComponents_linear = kernel_pca.fit_transform(standMasch_df)
#Store results
principalDf_linear = pd.DataFrame(data = principalComponents_linear, columns = ['principal component 1', 'principal component 2'])

#Plot results
fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('Principal Component 1', fontsize = 15)
ax.set_ylabel('Principal Component 2', fontsize = 15)
ax.set_title('2 component linear kernel PCA', fontsize = 20)

ax.scatter(principalDf_linear['principal component 1'],principalDf_linear['principal component 2'])
ax.grid()

# Polynomial Kernel

#apply polynomial kernel and then perform PCA
kernel_pca = KernelPCA(n_components=2, kernel="poly", gamma=None, fit_inverse_transform=True, alpha=0.1)
principalComponents_poly = kernel_pca.fit_transform(standMasch_df)
#Store results
principalDf_poly = pd.DataFrame(data = principalComponents_poly, columns = ['principal component 1', 'principal component 2'])

#Plot results
fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('Principal Component 1', fontsize = 15)
ax.set_ylabel('Principal Component 2', fontsize = 15)
ax.set_title('2 component polynomial kernel PCA', fontsize = 20)

ax.scatter(principalDf_poly['principal component 1'],principalDf_poly['principal component 2'])
ax.grid()

#TODO: use PCA on entire test data not just practice data, so macht das Geoclustern keinen Sinn
# Plot first cluster geodata
fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('latitude', fontsize = 15)
ax.set_ylabel('longitude', fontsize = 15)
ax.set_title('Cluster 1 Geo Data', fontsize = 20)

ax.scatter(cluster1df['latitude'],cluster1df['longitude'],s = 0.1)
ax.grid()

# Singular Value Decomposition

svd = TruncatedSVD(n_components=2, n_iter=7, random_state=42)
svd.fit(dfMaschen)

print(svd.explained_variance_ratio_)
print(svd.explained_variance_ratio_.sum())
print(svd.singular_values_)

# Non linear projections
